improve the project structure
This commit improves the structure of the entire project, especially for
mirrorlist. Four modules state, mutations, config and update are
created.

From now on, mirrorlist no longer updates itself. Instead, it reads
`state` which is updated by `mutations` triggered by `update`.
