+++ src/config.js
@@ -1,0 +1,5 @@
+export default {
+    statusUrl: 'http://localhost:7001/lug/v1/manager',
+    mirrorBaseUrl: 'http://localhost:8000/', // don't miss the terminating slash
+    updateInterval: 30000,
+};
