+++ src/update.js
@@ -1,0 +1,13 @@
+import config from './config';
+import mutations from './mutations';
+
+export default {
+    updateMirrorListRoutine() {
+        mutations.updateMirrorList();
+        setTimeout(this.updateMirrorListRoutine, config.updateInterval);
+    },
+
+    start() {
+        this.updateMirrorListRoutine();
+    },
+};
